Fixed few issues with YAML and prompt
